+++ .gitignore
@@ -2,3 +2,4 @@
 *.pdf
 Gemfile.lock
 coverage
+.byebug_history
