+++ lib/br_invoices_pdf/cfe/parser/cpf.rb
@@ -1,0 +1,13 @@
+module BrInvoicesPdf
+  module Cfe
+    module Parser
+      class Cpf
+        include BaseParser
+
+        def execute(xml)
+          locate_element(xml, 'infCFe/dest/CPF')
+        end
+      end
+    end
+  end
+end
