+++ lib/br_invoices_pdf/cfe/parser/document_number.rb
@@ -1,0 +1,13 @@
+module BrInvoicesPdf
+  module Cfe
+    module Parser
+      class DocumentNumber
+        include BaseParser
+
+        def execute(xml)
+          locate_element(xml, 'infCFe/ide/nCFe')
+        end
+      end
+    end
+  end
+end
