+++ lib/br_invoices_pdf/cfe/renderer/taxes_info.rb
@@ -1,0 +1,21 @@
+module BrInvoicesPdf
+  module Cfe
+    module Renderer
+      class TaxesInfo
+        include BaseRenderer
+
+        def execute(pdf, data)
+          box(pdf, [0, pdf.cursor], page_content_width(pdf)) do
+            pdf.text("Tributos\n\n", style: :italic)
+            value = format_currency(BigDecimal(data[:payment][:approximate_value_of_taxex]))
+            text = "Informação dos tributos totais incidentes (Lei Federal 12.741/2012):\n R$ #{value}\n\n"
+            pdf.text(text, align: :center)
+            pdf.text('SAT Número ' + data[:sat_params][:sat_number], align: :center)
+            time = data[:sat_params][:emission_date] + data[:sat_params][:emission_hour]
+            pdf.text(DateTime.parse(time).strftime('%d/%m/%Y %H:%M:%S'), align: :center)
+          end
+        end
+      end
+    end
+  end
+end
